backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
pydantic==1.8.2
kafka-python==2.0.2
requests==2.26.0
python-dotenv==0.19.0
prometheus-client==0.11.0
httpx==0.24.1